requirements.txt
defusedxml==0.6.0
django==2.0
#django-cors-header==3.2.1
djangorestframework==3.9.0
djangorestframework-xml==1.4.0
pytz==2019.3